requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
alembic>=1.13.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
joblib>=1.3.0
pyarrow>=14.0.1
psycopg2-pool>=1.1.0
matplotlib>=3.7.0
